+++ dockci/util.py
@@ -34,7 +34,7 @@ from yaml_model import ValidationError
 AUTH_TOKEN_EXPIRY = 36000  # 10 hours
 
 
-def request_fill(model_obj, fill_atts, data=None, save=True):
+def request_fill(model_obj, fill_atts, accept_blank=(), data=None, save=True):
     """
     Fill given model attrs from a POST request (and ignore other requests).
     Will save only if the save flag is True
@@ -44,7 +44,7 @@ def request_fill(model_obj, fill_atts, data=None, save=True):
 
     if request.method == 'POST':
         for att in fill_atts:
-            if att in data and data[att] != '':
+            if att in data and (data[att] != '' or att in accept_blank):
                 setattr(model_obj, att, data[att])
             elif att not in data:  # For check boxes
                 setattr(model_obj, att, None)
@@ -308,7 +308,7 @@ class FauxDockerLog(object):
     """
     A contextual logger to output JSON lines to a handle
     """
-    def __init__(self, handle):
+    def __init__(self, handle=None):
         self.handle = handle
         self.defaults = {}
 
@@ -332,9 +332,12 @@ class FauxDockerLog(object):
         Write a JSON line with kwargs, and defaults combined
         """
         with self.more_defaults(**kwargs):
-            self.handle.write(json.dumps(self.defaults).encode())
-            self.handle.write('\n'.encode())
-            self.handle.flush()
+            line = ('%s\n' % json.dumps(self.defaults)).encode()
+            if self.handle is None:
+                yield line
+            else:
+                self.handle.write(line)
+                self.handle.flush()
 
 
 def tokengetter_for(oauth_app):