+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Send HTTP 204 to GitHub ref deletions #322
 - Add GitLab project type #323
 - Push job status for GitLab projects #323
+- Registry login #328
+- Errors in Docker JSON stream fail stage automatically #328
 
 ### v0.0.8 (Breaking)
 - **BREAKING** Replace data store with PostgreSQL/SQLAlchemy #260